extract errors and ops from table
